requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pyotp==2.9.0
python-dotenv==1.0.1
pydantic==2.10.3
